[feat] 대시보드 컴포넌트 분리
